+++ NEWS.md
@@ -1,5 +1,7 @@
 # cranlogs 2.1.1
 
+* New function cranlogs_badge() to construct the Markdown string of a badge for the README.
+
 * Update the definition of the argument values "last-day", "last-week" and "last-month" to make them consistent with the database: "last-day" is the last day for which data is available, "last-week" 6 days prior to that last day with data, "last-month" 29 days prior to that last day with data.
 
 * Drop row names after reordering which avoids having meaningless row names (@patperry, #23)